requirements.txt
fastapi
pydantic
uvicorn
supabase
python-jose[cryptography] 
passlib[bcrypt]
python-multipart